extract build module
