+++ scripts/dirname.mjs
@@ -1,0 +1,6 @@
+import path from 'node:path';
+import { fileURLToPath } from 'node:url';
+
+const __filename = fileURLToPath(import.meta.url);
+export const __dirname = path.dirname(__filename);
+export const wd = path.resolve(__dirname, '..');
